Fix -Wsign-compare warnings in MBString

+++ ext/mbstring/php_mbregex.c
@@ -699,7 +699,7 @@ mb_regex_groups_iter(const OnigUChar* name, const OnigUChar* name_end, int ngrou
 	gn = onig_name_to_backref_number(reg, name, name_end, args->region);
 	beg = args->region->beg[gn];
 	end = args->region->end[gn];
-	if (beg >= 0 && beg < end && end <= args->search_len) {
+	if (beg >= 0 && beg < end && (size_t)end <= args->search_len) {
 		add_assoc_stringl_ex(args->groups, (char *)name, name_end - name, &args->search_str[beg], end - beg);
 	} else {
 		add_assoc_bool_ex(args->groups, (char *)name, name_end - name, 0);
@@ -1428,7 +1428,7 @@ _php_mb_regex_ereg_search_exec(INTERNAL_FUNCTION_PARAMETERS, int mode)
 			for (i = 0; i < n; i++) {
 				beg = MBREX(search_regs)->beg[i];
 				end = MBREX(search_regs)->end[i];
-				if (beg >= 0 && beg <= end && end <= len) {
+				if (beg >= 0 && beg <= end && (size_t)end <= len) {
 					add_index_stringl(return_value, i, (char *)&str[beg], end - beg);
 				} else {
 					add_index_bool(return_value, i, 0);
@@ -1449,7 +1449,7 @@ _php_mb_regex_ereg_search_exec(INTERNAL_FUNCTION_PARAMETERS, int mode)
 			break;
 		}
 		end = MBREX(search_regs)->end[0];
-		if (pos <= end) {
+		if (pos <= (size_t)end) {
 			MBREX(search_pos) = end;
 		} else {
 			MBREX(search_pos) = pos + 1;
@@ -1568,7 +1568,7 @@ PHP_FUNCTION(mb_ereg_search_getregs)
 		for (i = 0; i < n; i++) {
 			beg = MBREX(search_regs)->beg[i];
 			end = MBREX(search_regs)->end[i];
-			if (beg >= 0 && beg <= end && end <= len) {
+			if (beg >= 0 && beg <= end && (size_t)end <= len) {
 				add_index_stringl(return_value, i, (char *)&str[beg], end - beg);
 			} else {
 				add_index_bool(return_value, i, 0);
